fix: 🐛 symlink scripts

+++ scripts/symlink-binaries.js
@@ -5,11 +5,11 @@ const {
 } = require('../lerna.json');
 const glob = require('glob');
 const _ = require('lodash');
+const { program } = require('commander');
 
-const symlink = () => {
+const symlink = (action = 'create') => {
     const BASE_DIR = path.resolve(__dirname, '..');
     const PROCESS_DIR = path.dirname(process.execPath);
-    const action = process.argv[2] || 'create';
 
     const packagePaths = packages.reduce((result, currentPattern) => {
         return result.concat(glob.sync(currentPattern, {
@@ -31,9 +31,15 @@ const symlink = () => {
                     continue;
                 }
 
+                fs.chmodSync(symlinkPathname, 777);
                 console.log('[SYMLINK][CREATE]', `${targetPathname} -> ${symlinkPathname}`);
                 fs.symlinkSync(targetPathname, symlinkPathname);
             } else if (action === 'remove') {
+                if (!fs.existsSync(symlinkPathname)) {
+                    console.log('[SYMLINK][REMOVE] skip', `${targetPathname} -> ${symlinkPathname}`);
+                    continue;
+                }
+
                 console.log('[SYMLINK][REMOVE]', symlinkPathname);
                 fs.removeSync(symlinkPathname);
             }
@@ -41,4 +47,10 @@ const symlink = () => {
     }
 };
 
-symlink();
+program
+    .option('-a, --action <action>', 'link binary action')
+    .action(async ({ action }) => {
+        symlink(action);
+    });
+
+program.parse(process.argv);
